Reworking the definition named constants

+++ Library/autoload.php
@@ -1,34 +1,31 @@
 <?php
 
-if (!defined('__EXECUTION_ACCESS_RESTRICTION__')) {
-  exit('No direct script access allowed');
-}
-
 /**
  * Finds and load a class from its full name, e.g. namespace + class name.
- * It will check if the file exists in the <i>__ROOT__</i> directory.
+ * It will check if the file exists in the <i>FrameworkConstants_RootDir</i> directory.
  * 
  * If it does, it will load the class to be used.
  * Otherwise, it will throw an exception.
  * 
- * @param string $className
+ * @param string $className : the full name, e.g. namespace + class name
  */
 function autoload($className) {
-  $file = __ROOT__ . str_replace('\\', '/', $className) . '.php';
-  if (file_exists($file)) {
-    try {
-      $calculateExeTime = defined("__ENABLE_BENCHMARK__") ? __ENABLE_BENCHMARK__ : FALSE;
-      if($calculateExeTime) { $start = microtime(TRUE); }
+  $file = FrameworkConstants_RootDir . str_replace('\\', '/', $className) . '.php';
+  $calculateExeTime = defined(FrameworkConstants_EnableBenchmark) ? FrameworkConstants_EnableBenchmark : FALSE;
+  try {
+    $fileExists = file_exists($file);
+    if ($fileExists && $calculateExeTime) {
+      $start = microtime(TRUE);
       require_once $file;
-      if($calculateExeTime) {
-        $exeTime = (microtime(TRUE) - $start) * 1000;
-        echo "loading class : $className : $exeTime ms<br />";
-      }
-    } catch (Exception $exc) {
-      echo "<!--" . $exc->getMessage() . "-->";
+      $exeTime = (microtime(TRUE) - $start) * 1000;
+      echo "loading class : $className : $exeTime ms<br />";
+    } elseif ($fileExists) {
+      require_once $file;
+    } else {
+      throw new ErrorException("Class not found => " . $file);
     }
-  } else {
-    throw new ErrorException("Class not found => " . $file);
+  } catch (Exception $exc) {
+    echo "<!--" . $exc->getMessage() . "-->";
   }
 }
